requirements_for_tests.txt
-r requirements.txt
nose==1.3.4
PyHamcrest==1.8.1
